Finalizes Conversion

+++ cmd/aasregistryservice/main.go
@@ -81,7 +81,8 @@ func runServer(ctx context.Context, configPath string, databaseSchema string) er
 
 	smDatabase, err := aasregistrydatabase.NewPostgreSQLAASRegistryDatabase(
 		dsn,
-		cfg.Postgres.MaxOpenConnections,
+		//nolint:gosec // configured value is bounded by deployment configuration
+		int32(cfg.Postgres.MaxOpenConnections),
 		cfg.Postgres.MaxIdleConnections,
 		cfg.Postgres.ConnMaxLifetimeMinutes,
 		cfg.Server.CacheEnabled,
